# backend-spring/target/generated-sources/openapi/src/main/java/org/example/generated/api/TasksApi.java
/*
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech) (7.15.0).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
package org.example.generated.api;

import org.example.generated.model.CreateTaskRequest;
import org.example.generated.model.Task;
import org.example.generated.model.UpdateTaskStatusRequest;
import io.swagger.v3.oas.annotations.ExternalDocumentation;
import io.swagger.v3.oas.annotations.Operation;
import io.swagger.v3.oas.annotations.Parameter;
import io.swagger.v3.oas.annotations.Parameters;
import io.swagger.v3.oas.annotations.media.ArraySchema;
import io.swagger.v3.oas.annotations.media.Content;
import io.swagger.v3.oas.annotations.media.Schema;
import io.swagger.v3.oas.annotations.responses.ApiResponse;
import io.swagger.v3.oas.annotations.security.SecurityRequirement;
import io.swagger.v3.oas.annotations.tags.Tag;
import io.swagger.v3.oas.annotations.enums.ParameterIn;
import io.swagger.v3.oas.annotations.media.ExampleObject;
import org.springframework.http.HttpStatus;
import org.springframework.http.MediaType;
import org.springframework.lang.Nullable;
import org.springframework.http.ResponseEntity;
import org.springframework.validation.annotation.Validated;
import org.springframework.web.bind.annotation.*;
import org.springframework.web.context.request.NativeWebRequest;
import org.springframework.web.multipart.MultipartFile;

import jakarta.validation.Valid;
import jakarta.validation.constraints.*;
import java.util.List;
import java.util.Map;
import java.util.Optional;
import jakarta.annotation.Generated;

@Generated(value = "org.openapitools.codegen.languages.SpringCodegen", date = "2026-03-22T20:37:12.681241+01:00[Europe/Berlin]", comments = "Generator version: 7.15.0")
@Validated
@Tag(name = "Tasks", description = "the Tasks API")
public interface TasksApi {

    default Optional<NativeWebRequest> getRequest() {
        return Optional.empty();
    }

    public static final String PATH_CREATE_TASK = "/tasks";
    /**
     * POST /tasks : Create a task
     *
     * @param createTaskRequest  (required)
     * @return Task created (status code 201)
     */
    @Operation(
        operationId = "createTask",
        summary = "Create a task",
        tags = { "Tasks" },
        responses = {
            @ApiResponse(responseCode = "201", description = "Task created", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = Task.class))
            })
        }
    )
    @RequestMapping(
        method = RequestMethod.POST,
        value = TasksApi.PATH_CREATE_TASK,
        produces = { "application/json" },
        consumes = { "application/json" }
    )
    
    default ResponseEntity<Task> createTask(
        @Parameter(name = "CreateTaskRequest", description = "", required = true) @Valid @RequestBody CreateTaskRequest createTaskRequest
    ) {
        getRequest().ifPresent(request -> {
            for (MediaType mediaType: MediaType.parseMediaTypes(request.getHeader("Accept"))) {
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"description\" : \"description\", \"id\" : \"id\", \"title\" : \"title\", \"status\" : \"OPEN\" }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
            }
        });
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }


    public static final String PATH_GET_TASK_BY_ID = "/tasks/{taskId}";
    /**
     * GET /tasks/{taskId} : Get task by id
     *
     * @param taskId  (required)
     * @return Task found (status code 200)
     *         or Task not found (status code 404)
     */
    @Operation(
        operationId = "getTaskById",
        summary = "Get task by id",
        tags = { "Tasks" },
        responses = {
            @ApiResponse(responseCode = "200", description = "Task found", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = Task.class))
            }),
            @ApiResponse(responseCode = "404", description = "Task not found")
        }
    )
    @RequestMapping(
        method = RequestMethod.GET,
        value = TasksApi.PATH_GET_TASK_BY_ID,
        produces = { "application/json" }
    )
    
    default ResponseEntity<Task> getTaskById(
        @Parameter(name = "taskId", description = "", required = true, in = ParameterIn.PATH) @PathVariable("taskId") String taskId
    ) {
        getRequest().ifPresent(request -> {
            for (MediaType mediaType: MediaType.parseMediaTypes(request.getHeader("Accept"))) {
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"description\" : \"description\", \"id\" : \"id\", \"title\" : \"title\", \"status\" : \"OPEN\" }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
            }
        });
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }


    public static final String PATH_GET_TASKS = "/tasks";
    /**
     * GET /tasks : Get all tasks
     *
     * @return List of tasks (status code 200)
     */
    @Operation(
        operationId = "getTasks",
        summary = "Get all tasks",
        tags = { "Tasks" },
        responses = {
            @ApiResponse(responseCode = "200", description = "List of tasks", content = {
                @Content(mediaType = "application/json", array = @ArraySchema(schema = @Schema(implementation = Task.class)))
            })
        }
    )
    @RequestMapping(
        method = RequestMethod.GET,
        value = TasksApi.PATH_GET_TASKS,
        produces = { "application/json" }
    )
    
    default ResponseEntity<List<Task>> getTasks(
        
    ) {
        getRequest().ifPresent(request -> {
            for (MediaType mediaType: MediaType.parseMediaTypes(request.getHeader("Accept"))) {
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "[ { \"description\" : \"description\", \"id\" : \"id\", \"title\" : \"title\", \"status\" : \"OPEN\" }, { \"description\" : \"description\", \"id\" : \"id\", \"title\" : \"title\", \"status\" : \"OPEN\" } ]";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
            }
        });
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }


    public static final String PATH_UPDATE_TASK_STATUS = "/tasks/{taskId}/status";
    /**
     * PATCH /tasks/{taskId}/status : Update task status
     *
     * @param taskId  (required)
     * @param updateTaskStatusRequest  (required)
     * @return Updated task (status code 200)
     */
    @Operation(
        operationId = "updateTaskStatus",
        summary = "Update task status",
        tags = { "Tasks" },
        responses = {
            @ApiResponse(responseCode = "200", description = "Updated task", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = Task.class))
            })
        }
    )
    @RequestMapping(
        method = RequestMethod.PATCH,
        value = TasksApi.PATH_UPDATE_TASK_STATUS,
        produces = { "application/json" },
        consumes = { "application/json" }
    )
    
    default ResponseEntity<Task> updateTaskStatus(
        @Parameter(name = "taskId", description = "", required = true, in = ParameterIn.PATH) @PathVariable("taskId") String taskId,
        @Parameter(name = "UpdateTaskStatusRequest", description = "", required = true) @Valid @RequestBody UpdateTaskStatusRequest updateTaskStatusRequest
    ) {
        getRequest().ifPresent(request -> {
            for (MediaType mediaType: MediaType.parseMediaTypes(request.getHeader("Accept"))) {
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"description\" : \"description\", \"id\" : \"id\", \"title\" : \"title\", \"status\" : \"OPEN\" }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
            }
        });
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }

}